Create Bot class
